+++ lib/main.ts
@@ -1,7 +1,11 @@
 export { result } from "./primitives/result.ts";
 export { zero } from "./primitives/zero.ts";
+export { take } from "./primitives/take.ts";
+export { takeTwo } from "./primitives/takeTwo.ts";
 
 export { any } from "./combinators/alternation/any.ts";
+export { defaulted } from "./combinators/choice/defaulted.ts";
+export { optional } from "./combinators/choice/optional.ts";
 export { or } from "./combinators/choice/or.ts";
 export { filter } from "./combinators/filtering/filter.ts";
 export { foldL } from "./combinators/iteration/foldL.ts";
@@ -18,3 +22,28 @@ export { lazy } from "./combinators/lazy-evaluation/lazy.ts";
 export { memoize } from "./combinators/lazy-evaluation/memoize.ts";
 export { bracket } from "./combinators/sequencing/bracket.ts";
 export { and } from "./combinators/sequencing/and.ts";
+export { first } from "./combinators/sequencing/first.ts";
+export { last } from "./combinators/sequencing/last.ts";
+
+export { alphaNums } from "./tokens/alphaNums.ts";
+export { decimal } from "./tokens/decimal.ts";
+export { digit } from "./tokens/digit.ts";
+export { identifier } from "./tokens/identifier.ts";
+export { integer } from "./tokens/integer.ts";
+export { letter } from "./tokens/letter.ts";
+export { letters } from "./tokens/letters.ts";
+export { listOf } from "./tokens/listOf.ts";
+export { listOfInts } from "./tokens/listOfInts.ts";
+export { literal } from "./tokens/literal.ts";
+export { lower } from "./tokens/lower.ts";
+export { natural } from "./tokens/natural.ts";
+export { newline } from "./tokens/newline.ts";
+export { number } from "./tokens/number.ts";
+export { regex } from "./tokens/regex.ts";
+export { regexPredicate } from "./tokens/regexPredicate.ts";
+export { spaces } from "./tokens/spaces.ts";
+export { token } from "./tokens/token.ts";
+export { upper } from "./tokens/upper.ts";
+export { whitespace } from "./tokens/whitespace.ts";
+export { whitespaces } from "./tokens/whitespaces.ts";
+export { whitespaces1 } from "./tokens/whitespaces1.ts";
