+++ lib/tokens/alphaNums.ts
@@ -1,0 +1,9 @@
+import type { Parser } from "../../src/index.ts";
+import { regex } from "./regex.ts";
+
+/**
+ * Parses alphanumeric characters (0 or more)
+ *
+ * * Regex: /^\w*\/
+ */
+export const alphaNums: Parser<string> = regex(/^\w*/);
