+++ lib/tokens/spaces.ts
@@ -1,0 +1,7 @@
+import type { Parser } from "../../src/index.ts";
+import { regex } from "./regex.ts";
+
+/**
+ * Parses the space character (0 or more)
+ */
+export const spaces: Parser<string> = regex(/^ */);
