+++ src/main.ts
@@ -1,23 +1,51 @@
 export { createParser, Parser } from "./index.ts";
 
 export {
+  alphaNums,
   and,
   any,
   bracket,
+  decimal,
+  defaulted,
+  digit,
   filter,
+  first,
   foldL,
   foldL1,
   foldR,
   foldR1,
+  identifier,
+  integer,
   iterate,
+  last,
   lazy,
+  letter,
+  letters,
+  listOf,
+  listOfInts,
+  literal,
+  lower,
   many,
   many1,
   memoize,
+  natural,
+  newline,
+  number,
+  optional,
   or,
+  regex,
+  regexPredicate,
   repeat,
   result,
   sepBy,
   sepBy1,
+  spaces,
+  take,
+  takeTwo,
+  token,
+  upper,
+  whitespace,
+  whitespaces,
+  whitespaces1,
   zero,
 } from "../lib/main.ts";
